% Generated by build script
\name{ibge_metadata}
\alias{ibge_metadata}
\title{Get metadata for an IBGE aggregate}
\usage{
ibge_metadata()
}
\arguments{
\item{aggregate_id}{Aggregate table ID (character or integer, e.g. "4714")}
}
\value{tibble: variable_id, variable_name, unit}
\description{Get metadata for an IBGE aggregate}
